shit auth for clean download folder

+++ main.py
@@ -1,11 +1,15 @@
-from flask import Flask, request, render_template, jsonify, send_file
+from flask import Flask, request, render_template, jsonify, send_file, redirect
 import youtube_dl
-import os.path
 import os
 import threading
 import time
+import random
 
 app = Flask(__name__)
+
+@app.errorhandler(404)
+def page_not_found(e):
+    return redirect("/", code=302)
 
 @app.route('/', methods=['GET'])
 def index():
@@ -28,13 +32,36 @@ def convert():
 def download(filename):
     if(os.path.isfile('./download/' + filename)):
         thread = threading.Thread(target=delete_file, kwargs={
-            'filename': filename
+            'filename': './download/' + filename
         })
         thread.start()
 
         return send_file('./download/' + filename, as_attachment=True), 200
 
     return 'Arquivo Não encontrado', 404
+
+@app.route('/randomword', methods=['GET'])
+def randomword():
+    word_list = open('wordlist.txt', 'r').read().split("\n")
+    word = random.choice(word_list)
+    open('./temp/word', 'w').write(word)
+
+    delete_word = threading.Thread(target=delete_file, kwargs={
+        'filename': './temp/word',
+        'time_wait': 30
+    })
+    delete_word.start()
+
+    return word
+
+@app.route('/maintenance', methods=['POST'])
+def maintenance():
+    auth = request.headers.get('auth-code')
+    if auth is not None:
+        if(os.path.isfile('./temp/word') and auth == open('./temp/word', 'r').read()):
+            clean_download_folder()
+            return open('./static/plastic_love.txt', 'r').read().replace("\n", "<br>"), 200
+    return redirect("/", code=302)
 
 def convert_mp3(video_link):
     try:
@@ -63,17 +90,19 @@ def convert_mp3(video_link):
     except:
         return False
 
-def delete_file(filename):
-    time.sleep(60)
-    os.remove('./download/' + filename)
+def delete_file(filename, time_wait = 60):
+    time.sleep(time_wait)
+    os.remove(filename)
 
 def clean_download_folder():
-    args = ['./download']
-    for item in args:
+    paths = ['./download']
+    for path in paths:
         i = 0 
-        for p, _, files in os.walk(os.path.abspath(item)):
+        for p, _, files in os.walk(os.path.abspath(path)):
             if(i != 0): # dont delete the root download path
                 os.rmdir(p)
             for file in files:
+                if file == '.gitkeep':
+                    continue
                 os.remove(os.path.join(p, file))
             i = i + 1
